requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
kubernetes==29.0.0
python-dotenv==1.0.1
email-validator==2.1.0.post1
cryptography==42.0.5 